Remove metadata todos, update entity bounding box sizes, move EvokerFangs to objects
