+++ src/main/java/net/glowstone/entity/monster/GlowElderGuardian.java
@@ -8,7 +8,7 @@ import org.bukkit.entity.EntityType;
 public class GlowElderGuardian extends GlowGuardian implements ElderGuardian {
     public GlowElderGuardian(Location loc) {
         super(loc, EntityType.ELDER_GUARDIAN, 80);
-        setBoundingBox(0.85, 0.85);
+        setBoundingBox(1.9975, 1.9975);
     }
 
     @Override
