+++ src/main/java/net/glowstone/entity/monster/GlowEvoker.java
@@ -14,6 +14,7 @@ public class GlowEvoker extends GlowMonster implements Evoker {
     public GlowEvoker(Location loc) {
         super(loc, EntityType.EVOKER, 24);
         metadata.set(MetadataIndex.EVOKER_SPELL, (byte) Spell.NONE.ordinal());
+        setBoundingBox(0.6, 1.95);
     }
 
     @Override
